+++ apps/gateway/src/utils/aws.ts
